+++ lib/terraspace/dependency/helper/base.rb
@@ -1,0 +1,7 @@
+module Terraspace::Dependency::Helper
+  class Base
+    def initialize(mod, identifier, options)
+      @mod, @identifier, @options = mod, identifier, options
+    end
+  end
+end